Fix bug in subroutine call
If we called a subroutine with a &rest argument and did not provide any
arguments to fill it, then the logic would insert a nil value there. That meant
that when we created the &rest slice it would be a slice of 1 nil object instead
of an empty slice. We fix this by changing the logic so an extra arg is only
pushed in the lisp object case.