Fix bug with Option types in defun signatures not being handled correctly
The issue is that there is a blanket impl `From<T> for Option<T>` and an impl
`TryFrom<T> for U where T: Into<U>`. So it was automatically converting option
types without checking if it was nil. So we defined a custom function
try_from_option that will handle this correctly in defun.