Remove byte code compiling until uninterned symbol bug fixed
Right now all symbols are interned. Meaning that if two symbols have the same
printed representation, they are the same symbol. However this means that
situtations like https://debbugs.gnu.org/cgi/bugreport.cgi?bug=43601 are real
bugs right now and not just a source of confusion. This is leading to cycles in
the data structures that cause a stack overflow (since we have no cycle detection).