Make Object covariant
This just makes things so much easier to work with. But that means we have to
remove all lifetimes in the objects and only have references to static types.
This required some extra work, but we now have Gc managed types that don't need
lifetime parameters. I think these are sound, but will need to think about it more.