Fix non-determinism bug in parsing DEFVAR_BOOL statements
fixes #27

The issue was that rust parses the project files in parallel and so the order
that we find "defvar" forms in non-deterministic. If we find a DEFVAR_BOOL form
before we find `BYTE_BOOLEAN_VARS`, then it will try and push to that variable
before it exists. We fix this by special casing `BYTE_BOOLEAN_VAR` and moving it
to the end of the init.