Use LazyLock for static variable initialization
LazyLock is the unstable version of the Lazy API for the OnceLock
struct. It's in the process of stabilization, but still guarded by the
`lazy_cell` feature.

Using LazyLock is more ergonomic, as it allows the callers to assume
that the contents will be initialized when called, and they can call the
methods of the contents directly (in this case, `Mutex`), instead of
`get`ting the OnceLock first and then using the methods.

This commit includes the change on the INTERNED_SYMBOLS static on the
build.rs file, which we no longer need to force initialization, since it
will be initialized lazily once first used, by definition of the
LazyLock API.