Finish bootstrapping to bytecomp with Emacs 29 elisp
We had to hack a lot of things to get to this point, but it now loads all the
lisp without errors. However we are not actually using the bytecompiled
functions. Doing so results in elisp byte-compile error that we need to debug.
But this at least gets us close to the point we had with the Emacs 27 elisp.