Use std OnceLock instead of lazy_static (#33)
* Use OnceLock instead of lazy_static

once_cell has the benefit of being macroless (easier to debug) plus the
pro of being introduced into the standard library. There are still some
unstable features for once_cell lazy usage, say, for LazyLock (makes
easier initializing the statics defined), though the API was introduced
in 1.70.0.

Instead of using the lazy_static macro and then creating the static ref
definitions inside, we can just define a OnceCell and then initialize it.

* Use LazyLock for static variable initialization

LazyLock is the unstable version of the Lazy API for the OnceLock
struct. It's in the process of stabilization, but still guarded by the
`lazy_cell` feature.

Using LazyLock is more ergonomic, as it allows the callers to assume
that the contents will be initialized when called, and they can call the
methods of the contents directly (in this case, `Mutex`), instead of
`get`ting the OnceLock first and then using the methods.

This commit includes the change on the INTERNED_SYMBOLS static on the
build.rs file, which we no longer need to force initialization, since it
will be initialized lazily once first used, by definition of the
LazyLock API.

* Remove lazy_static references on tests

We no longer need to initialize it, as by nature the LazyLock API will
ensure the content is initialized on the first usage, so the consumers
can just call the normal content methods.

* Use the stable OnceLock and add TODO for LazyLock usage

Even though LazyLock is more ergonomic, we want to keep the code base as
stable for now. We use the `beta` channel to run the code. Once the
`LazyLock` API is stabilized, we can track that small migration work on
this issue:

See: #34