Document scheduled removal and replace master
Scheduled is a job that runs daily whenever there's a merge on nigthly,
with the hypothesis that it's possible that nightly builds could break
our crate. Since we are not using nightly (just for CI), I propose we
iterate on it. If we feel our CI breaks because of nightly builds, we
can add it back.

For sure we could consider adding it if we ever move our toolchain to nightly.