Deleting rust-toolchain.toml
Having `info: syncing channel updates for
'beta-x86_64-unknown-linux-gnu'` as the toolchain on the repo
is actually breaking CI. Let's see other crates and see what
they are doing.

- serde: no rust-toolchain.toml
- once_cell: no rust-toolchain.toml
- axum: no rust-toolchain.toml

My hypothesis for that is that you don't really need a
rust-toolchain.toml checked out on the repo, but rather we should
ignore it. If you want to use a rust toolchain great, so you avoid
having to run cargo with =+beta=, but no need to have it on version control.