Stability: Add CI to check, test and run safety checks (#35)
* Add initial files and document potential improvements

* Rename branches to `master`

Checked with the owner whether he has a preference on `master` compared
to `main`, as we see a general trend of Github repos choosing the later.

* Remove nostd workflow

We don't really support `nostd` enviroments and in all honesty, I don't
see how this project would be used on those.

* Add MIRIFLAGS as specified on README.org

* Document scheduled removal and replace master

Scheduled is a job that runs daily whenever there's a merge on nigthly,
with the hypothesis that it's possible that nightly builds could break
our crate. Since we are not using nightly (just for CI), I propose we
iterate on it. If we feel our CI breaks because of nightly builds, we
can add it back.

For sure we could consider adding it if we ever move our toolchain to nightly.

* Run fmt --check on the stable toolchain

* Enforce nightly on cargo doc

By having the rust toolchain point to beta, we are failing on some of
the CI jobs.

* Deleting rust-toolchain.toml

Having `info: syncing channel updates for
'beta-x86_64-unknown-linux-gnu'` as the toolchain on the repo
is actually breaking CI. Let's see other crates and see what
they are doing.

- serde: no rust-toolchain.toml
- once_cell: no rust-toolchain.toml
- axum: no rust-toolchain.toml

My hypothesis for that is that you don't really need a
rust-toolchain.toml checked out on the repo, but rather we should
ignore it. If you want to use a rust toolchain great, so you avoid
having to run cargo with =+beta=, but no need to have it on version control.

* WIP: Install the pango library before runnning clippy

* Install the gtk-3.0 system library

* Add the system libraries to check and msrv

* Install dependencies on satety and test workflows

* Remove loom and document possible improvements

* Remove ubuntu system on macos and windows

* Add codecov.io to possible improvements

* Test improvement with directed-minimal-versions

* Revert "Test improvement with directed-minimal-versions"

This reverts commit 60d38ff33a093da572619fbffd3c114c2e29aad2.

* Fix sanitizers not finding any library code

We don't really own any, we are a binary to run Emacs by loading its
elisp files, providing the low level functionality that emacs can't
implement in elisp.

* Remove windows-latest from os matrix

We don't currently support Windows (at least as we see from CI we are
panicking), but definitely something to consider for the future. Added
improvement to DOCS.md.

* Add improvement around doctests on binary crates