Modify CI to run directed-minimal-versions
`-Zdirected-minimal-versions` will run with the minimal versions of the
specified direct dependencies (of rune and all workspace crates), but
the maximum indirect dependencies. This means that we won't fail CI for
any issues our direct deps have with `-Zminimal-versions`.

Additionally, fixed some of the versions that were selected for syn,
quote and proc-macro2. It's recommended to use qualified versions, to
avoid issues with minimal versions.