Refactor fn_macros to rune-macros (#44)
* Refactor fn_macros to rune-macros

The reasoning behind this is that this is the start of the
modularization of the general crate, where I'm following the examples
structure of some famous crates: axum (library) and
ripgrep (binary). Axum has an `axum` preffix behind all the crates its
main packages uses, so thought it would be beneficial to do the same.

Also, moved rune-macros (renamed) inside `crates`, as it's a crate we
depend on, and we already had one there. It's a good practice to list
the members of the workspace directly, instead of using the wildcard, to
avoid accidential build inclusion.

* Add no_run to defun example failing doc test

* Swap no_run for ignore

* Fix windows build on rune-macros refactor

---------

Co-authored-by: Troy Hinckley <t.macman@gmail.com>