Publish rune docs to netlify (#50)
I tried different things and ended up choosing netlify. Netlify has a
very powerful build system and it's running on Ubuntu, but it did not
allow for apt-get installs. Since we need pango and gdk to run `cargo
doc` on the crate, moved to Github actions.

As part of the same PR, I fixed all the warnings on cargo doc, making
sure that the documentation is correctly published. The trick to deploy
cargo doc is to deploy `target/doc` and then pair that with a redirect
of the main page (configured rune-rs.netlify.app on my own account) to
`/rune` which (as the main crate) has the info about the subcrates as well.

Closes https://github.com/CeleritasCelery/rune/issues/48