Create initial rune-core crate with only hashmap
This is the first step towards migrating the core out of the rune to a
different crate: rune-core. The idea with separating the PRs that I do
for this refactoring is to make the whole process easier to review +
making sure there are no regressions on the refactors we do.

I started moving hashmap, which is typing that is used both in rune-core
and the main crate, hence why moved to rune-core. I'll move other
modules to rune-core little by little, as we get closer with PRs.