Add macros core module (#53)
On this module, we reexport all the macros that were previously defined
across the main crate. By reexporting under a module (rune_core::macros)
we avoid naming conflicts. An example of such conflict is the
rune::error module, since there were cases where we needed both the
`error!` macro and the `error` module.

By using `#[doc(hidden)]` on the private items, we avoid the
private (preffixed with __) to be added to the cargo doc documentation,
to avoid missusage of the macros.