Relax bounds on StackRoot
We don't need a 'static bound here because the stackroot is already bound the
original object and can't outlive it. Removing the 'static bounds let's us use
other lifetimes within a rooted struct then just 'static. But we also need to be
careful when using IntoRoot that we don't accidentally make something 'static
when it should not be.