Optimize function calls
We only want to do an allocation when we absolutely have to. So if the Env is
not mutably borrowed we just use the stack slice directly. If it mutably
borrowed, but the number of args is bounded then we can allocate them on the
stack. But if the args are unbounded we need to allocate.