Implement ArgSlice type
This type represents a `&[GcObj]` but doesn't actually hold the reference.
Instead the ArgSlice is passed to the stack to get the argument slice. This
let's us avoid allocations during calls which have variable numbers of
arguments.

Also fixed an unsoundness bug with Rt::bind_slice. We were changing the lifetime
of the slice itself, which could lead to use-after-free.