Split unibyte and multibyte strings into different types
99% of the time we are working with multibyte strings (str in rust parlance) But
having to always check that the string is valid pollutes the code. So instead we
are going to have a separate type for unibyte strings (byte slices). If there
are operations that are valid on both we will support those. However from the
elisp perspective nothing has changed. The type of both of these are "String".