Rework root macro with autoref specialization
We want to call IntoRoot if the type supports it. this let's us avoid hard to
understand lifetime errors when we pass a type that already has a lifetime. But
lots of types don't have an IntoRoot impl (and we don't really want to define
one). Using autoref specialization we can let the runtime decide for us.