Use slots to enable moving garbage collector
Initially the idea was to change `Rt` to contain an unsafe cell, since we need to
indicate to the type system that rooted values could be relocated. However I
[found
out](https://doc.rust-lang.org/std/cell/struct.UnsafeCell.html#memory-layout)
that even though UnsafeCell is repr(transparent) it disables niche optimizations
on the type which can change its size and alignment (i.e. an
`Option<UnsafeCell<T>>` is not same size as a `Option<T>`). This lead to a bunch of
hard to debug issues because we were using casts/transmutes to convert between T
and `Rt<T>`. We might have been able to make it work, but it would have been
painful and prone to UB if we got it wrong. So instead we added a new type
`Slot` that is a `UnsafeCell` wrapper around pointers that can move during
collection.