Implement casefiddle methods
This implements most of the casefiddle methods, with the exception of
*-region methods, as it seems that region manipulation isn't quite
implemented yet (though I could be wrong).  The implementation for
*-region ought to be simple anyway, just `kill-region` and `yank` with
`yank-transform-functions` set to the appropriate casefiddle method.

This provides unicode support via the standard library.  There may be
subtle differences between Rust's and Emacs' implementations, but those
can be addressed as they come up.  Complete support will require
implementing syntax tables to know whether a character makes a word or
symbol.