Introduce a new trait TracePtr to prevent immovable types being traced
If you just used `#[derive(Trace)]` you could end up accidentally including pointer types like
references or Gc<T> that would not be moved when GC is called. So we removed the implementation of
Trace from those types and implemented a new trait TracePtr. Now Trace is only implemented for Slot
types, which are the right types to include in other structs that will derive Trace.