Introduce ability to run Windows tests via Wine
Of course real testing should be done on Windows, but it is so painful
to develop Rune inside QEMU or even on Windows bare metal. TRAMP
doesn't work with Powershell or bash.exe from MingW32, so you have to
physically install Emacs on Windows inside QEMU and develop Rune this
way. It makes switching painful and very unproductive.

The provided way to run Windows tests via Wine is not ideal either,
but it works pretty well! Especially, when combined with changing the
target for rust-analyzer and hence eglot.